requirements.txt
# AI Image Generation Dependencies

# Core ML and Image Generation
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.20.0
safetensors>=0.3.0

# Hugging Face ecosystem
huggingface-hub>=0.16.0
datasets>=2.14.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0

# Ollama integration
ollama>=0.1.0
requests>=2.31.0
aiohttp>=3.8.0

# GUI Framework
PyQt6>=6.5.0
PyQt6-tools>=6.4.0

# Web Interface
gradio>=3.40.0
fastapi>=0.100.0
uvicorn>=0.23.0

# CLI and utilities
click>=8.1.0
typer>=0.9.0
tqdm>=4.65.0
rich>=13.5.0

# Configuration and data
pydantic>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
toml>=0.10.0

# Async support
asyncio-mqtt>=0.13.0
aiofiles>=23.0.0

# Development tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Optional: Advanced features
streamlit>=1.25.0  # Alternative web UI
wandb>=0.15.0     # Experiment tracking
tensorboard>=2.13.0  # Monitoring

# System utilities
psutil>=5.9.0
GPUtil>=1.4.0     # GPU monitoring
